refactor : 이미지 사이즈 관련 클래스 분리 및 테스트 코드 이동
